+++ scripts/pack.js
@@ -2,10 +2,10 @@
 
 const fs = require('fs')
 const path = require('path')
+const zlib = require('zlib')
 const rimraf = require('rimraf')
 const tar = require('tar-stream')
 const ZipStream = require('zip-stream')
-const zlib = require('zlib')
 const { version } = require('../package.json')
 
 const bin = path.resolve(__dirname, '../bin')
@@ -17,7 +17,7 @@ const os = (process.env.MATRIX_OS || 'linux,macos,windows').toLowerCase()
 const packToTarGz = (binary) => {
   const pack = tar.pack()
 
-  pack.entry({ name: binaryName, mode: 0755 }, binary)
+  pack.entry({ name: binaryName, mode: 755 }, binary)
   pack.finalize()
 
   return pack.pipe(zlib.createGzip())