+++ test/.eslintrc.js
@@ -1,0 +1,4 @@
+module.exports = {
+  plugins: ['jest'],
+  extends: ['plugin:jest/recommended', 'plugin:jest/style'],
+}