+++ test/preview.ts
@@ -1,8 +1,8 @@
 import path from 'path'
 import { ConvertType } from '../src/converter'
+import { CLIError } from '../src/error'
 import { File, FileType } from '../src/file'
 import { Preview, fileToURI } from '../src/preview'
-import { CLIError } from '../src/error'
 
 jest.mock('path')
 jest.setTimeout(15000)
@@ -36,7 +36,7 @@ describe('Preview', () => {
       instance.on('launch', launchEvent)
 
       await instance.open('about:blank')
-      expect(launchEvent).toBeCalledTimes(1)
+      expect(launchEvent).toHaveBeenCalledTimes(1)
     })
 
     it('emits opening event with location', async () => {
@@ -45,8 +45,8 @@ describe('Preview', () => {
       instance.on('opening', openingEvent)
 
       await instance.open('about:blank')
-      expect(openingEvent).toBeCalledTimes(1)
-      expect(openingEvent).toBeCalledWith('about:blank')
+      expect(openingEvent).toHaveBeenCalledTimes(1)
+      expect(openingEvent).toHaveBeenCalledWith('about:blank')
     })
 
     it('emits open event with window instance and location', async () => {
@@ -55,14 +55,14 @@ describe('Preview', () => {
       instance.on('open', openEvent)
 
       await instance.open('about:blank')
-      expect(openEvent).toBeCalledTimes(1)
+      expect(openEvent).toHaveBeenCalledTimes(1)
 
       const [win, location] = openEvent.mock.calls[0]
       expect(location).toBe('about:blank')
       expect(win.page.url()).toBe('about:blank')
     })
 
-    context('with constructor option about window size', () => {
+    describe('with constructor option about window size', () => {
       it('opens window that have specified window size', async () => {
         const instance = preview({ height: 400, width: 200 })
         const win = await instance.open('about:blank')
@@ -76,7 +76,7 @@ describe('Preview', () => {
       })
     })
 
-    context('when calling twice', () => {
+    describe('when calling twice', () => {
       it('opens 2 windows', async () => {
         const instance = preview()
         await instance.open('about:blank')
@@ -98,40 +98,43 @@ describe('Preview', () => {
         await instance.open('about:blank')
         await instance.open('about:blank')
 
-        expect(launchEvent).toBeCalledTimes(1)
-        expect(openingEvent).toBeCalledTimes(2)
-        expect(openEvent).toBeCalledTimes(2)
+        expect(launchEvent).toHaveBeenCalledTimes(1)
+        expect(openingEvent).toHaveBeenCalledTimes(2)
+        expect(openEvent).toHaveBeenCalledTimes(2)
       })
 
-      context('when opened window is closed', () => {
-        it('emits close event with closed window', async (done) => {
-          const instance = preview()
-          const closeEvent = jest.fn()
+      describe('when opened window is closed', () => {
+        it('emits close event with closed window', () =>
+          new Promise((done) =>
+            (async () => {
+              const instance = preview()
+              const closeEvent = jest.fn()
 
-          instance.on('close', closeEvent)
-          instance.on('exit', () => {
-            expect(closeEvent).toBeCalledTimes(2)
-            expect(closeEvent).toBeCalledWith(win)
-            expect(closeEvent).toBeCalledWith(win2)
-            done()
-          })
+              instance.on('close', closeEvent)
+              instance.on('exit', () => {
+                expect(closeEvent).toHaveBeenCalledTimes(2)
+                expect(closeEvent).toHaveBeenCalledWith(win)
+                expect(closeEvent).toHaveBeenCalledWith(win2)
+                done()
+              })
 
-          const win = await instance.open('about:blank')
-          const win2 = await instance.open('about:blank')
+              const win = await instance.open('about:blank')
+              const win2 = await instance.open('about:blank')
 
-          await win.close()
-          await win2.close()
-        })
+              await win.close()
+              await win2.close()
+            })()
+          ))
       })
     })
   })
 })
 
 describe('#fileToURI', () => {
-  context('with passing file', () => {
+  describe('with passing file', () => {
     const { posix, win32 } = path as any
 
-    context('in posix file system', () => {
+    describe('in posix file system', () => {
       it('returns file schema URI', () => {
         posix()
         expect(fileToURI(new File('/a/b/c'), ConvertType.html)).toBe(
@@ -140,7 +143,7 @@ describe('#fileToURI', () => {
       })
     })
 
-    context('in Windows file system', () => {
+    describe('in Windows file system', () => {
       it('returns file schema URI', () => {
         win32()
         expect(fileToURI(new File('c:\\abc'), ConvertType.html)).toBe(
@@ -150,7 +153,7 @@ describe('#fileToURI', () => {
     })
   })
 
-  context('with passing standard IO buffer', () => {
+  describe('with passing standard IO buffer', () => {
     const file = () => {
       const fileInstance = new File('')
 
@@ -176,7 +179,7 @@ describe('#fileToURI', () => {
       )
     })
 
-    context('when buffer is not ready', () => {
+    describe('when buffer is not ready', () => {
       it('throws CLIError', () =>
         expect(() => fileToURI(file(), ConvertType.html)).toThrow(CLIError))
     })